+++ Examples/Map/Sources/Constants.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2023 Wemap SAS. All rights reserved.
 //
 
-import CoreLocation
 import Foundation
 import WemapCoreSDK
 import WemapMapSDK
@@ -22,7 +21,7 @@ enum Constants {
     }
     
     static var polestarApiKey: String {
-        fatalError("Specify polestarApiKey and remove fatalError")
+        fatalError("Specify token and remove fatalError")
     }
     
     static var globalNavigationOptions: NavigationOptions {
@@ -30,5 +29,25 @@ enum Constants {
             itineraryOptions: .init(color: .cyan, projectionOptions: .init(width: 5, color: .lightGray)),
             userTrackingMode: .followWithHeading
         )
+    }
+}
+
+enum UIConstants {
+    
+    enum Delay {
+        static let short: TimeInterval = 5
+        static let long: TimeInterval = 10
+    }
+    
+    enum Inset {
+        static let top: CGFloat = -200
+        static let mid: CGFloat = -150
+    }
+}
+
+extension UserDefaults {
+    
+    static func double(forKey key: PreferencesKey, defaultValue: Double) -> Double {
+        UserDefaults.value(forKey: key) != nil ? standard.double(forKey: key.rawValue) : defaultValue
     }
 }