+++ Examples/Map/Sources/ViewControllers/NavigationViewController.swift
@@ -14,13 +14,14 @@ import WemapMapSDK
 
 final class NavigationViewController: MapViewController {
     
+    typealias Delay = UIConstants.Delay
+    typealias Inset = UIConstants.Inset
+    
     @IBOutlet var startNavigationButton: UIButton!
     @IBOutlet var stopNavigationButton: UIButton!
     @IBOutlet var startNavigationFromUserCreatedAnnotationsButton: UIButton!
     @IBOutlet var removeUserCreatedAnnotationsButton: UIButton!
     @IBOutlet var navigationInfo: UILabel!
-    
-    private let disposeBag = DisposeBag()
     
     private var userCreatedAnnotations: [MGLAnnotation] {
         map.annotations?
@@ -31,24 +32,17 @@ final class NavigationViewController: MapViewController {
         map.userLocationManager.locationSource as? LocationSourceSimulator
     }
     
+    private var navigationManager: NavigationManager { map.navigationManager }
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        map.navigationManager.delegate = self
+        navigationManager.delegate = self
+        pointOfInterestManager.delegate = self
         
         createLongPressGestureRecognizer()
         
         map.userTrackingMode = .followWithHeading
-        
-        // to see coordinate returned by location source
-        weak var previous: UIView?
-        map.userLocationManager
-            .coordinateUpdated
-            .subscribe(onNext: { [unowned self] in
-                previous?.removeFromSuperview()
-                previous = ToastHelper.showToast(message: "Location: \($0)", onView: view, bottomInset: -200)
-            })
-            .disposed(by: disposeBag)
         
         // this way you can specify user location indicator appearance
 //        map.userLocationManager.userLocationAnnotationViewStyle = .init(
@@ -62,8 +56,11 @@ final class NavigationViewController: MapViewController {
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         
-        ToastHelper.showToast(message: "Create 1 or 2 annotations by long press on the map to be able to start navigation. " +
-            "1 annotation to start from user location. 2 annotations to start from custom location", onView: view, hideDelay: 5)
+        ToastHelper.showToast(
+            message: "Create 1 or 2 annotations by long press on the map to be able to start navigation. " +
+                "1 annotation to start from user location. 2 annotations to start from custom location",
+            onView: view, hideDelay: Delay.short
+        )
     }
     
     @IBAction func closeTouched() {
@@ -87,81 +84,38 @@ final class NavigationViewController: MapViewController {
         let point = MGLPointAnnotation()
         point.coordinate = coord
         point.title = "user-created"
-        point.subtitle = "\(map.buildingManager.focusedBuilding?.activeLevel.id ?? 0.0)"
+        point.subtitle = "\(focusedBuilding?.activeLevel.id ?? 0.0)"
         map.addAnnotation(point)
-        updateStartNavigationButtons()
-        removeUserCreatedAnnotationsButton.isEnabled = true
+        updateUI()
     }
     
     @IBAction func startNavigationFromUserLocation() {
-        startNavigationButton.isEnabled = false
-        startNavigationFromUserCreatedAnnotationsButton.isEnabled = false
-        
         let to = userCreatedAnnotations.first!
-        
         let destination = Coordinate(coordinate2D: to.coordinate, level: Float(to.subtitle!!)!)
         
-        map.navigationManager
-            .startNavigation(to: destination, options: Constants.globalNavigationOptions)
-            .subscribe(
-                onSuccess: { [unowned self] itinerary in
-                    debugPrint("Navigation started successfully")
-                    stopNavigationButton.isEnabled = true
-                    simulator?.setItinerary(itinerary)
-                },
-                onFailure: { [unowned self] error in
-                    debugPrint("Failed to start navigation with error - \(error)")
-                    updateStartNavigationButtons()
-                    ToastHelper.showToast(
-                        message: "Failed to start navigation from user position to - \(destination) with error - \(error)",
-                        onView: view,
-                        hideDelay: 10
-                    )
-                }
-            ).disposed(by: disposeBag)
+        startNavigation(origin: nil, destination: destination)
     }
     
     @IBAction func stopNavigation() {
-        let result = map.navigationManager.stopNavigation()
+        let result = navigationManager.stopNavigation()
         switch result {
         case .success:
-            updateUIForNavigationStop()
+            simulator?.reset()
+            stopNavigationButton.isEnabled = false
+            updateUI()
         case let .failure(error):
-            debugPrint("Failed to stop navigation with error - \(error)")
-            ToastHelper.showToast(
-                message: "Failed to stop navigation with error - \(error)",
-                onView: view,
-                hideDelay: 10
-            )
+            ToastHelper.showToast(message: "Failed to stop navigation with error - \(error)", onView: view, hideDelay: Delay.long)
         }
-    }
-    
-    private func updateUIForNavigationStop() {
-        updateStartNavigationButtons()
-        simulator?.reset()
-        navigationInfo.isHidden = true
-        navigationInfo.text = nil
-    }
-    
-    private func getLevelFromAnnotation(_ annotation: MGLAnnotation) -> [Float] {
-        guard let building = map.buildingManager.focusedBuilding else {
-            debugPrint("Failed to retrieve focused building. Can't check if annotation is indoor or outdoor")
-            return []
-        }
-                
-        return building.boundingBox.contains(annotation.coordinate) ? [Float(annotation.subtitle!!)!] : []
     }
     
     @IBAction func startNavigationFromUserCreatedAnnotations() {
-        startNavigationButton.isEnabled = false
-        startNavigationFromUserCreatedAnnotationsButton.isEnabled = false
         
         let from = userCreatedAnnotations[0]
         let to = userCreatedAnnotations[1]
         
         let fromLevels = getLevelFromAnnotation(from)
         let toLevels = getLevelFromAnnotation(to)
-
+        
         let origin, destination: Coordinate
         if locationSourceType != .polestarEmulator {
             origin = Coordinate(coordinate2D: from.coordinate, levels: fromLevels)
@@ -179,43 +133,71 @@ final class NavigationViewController: MapViewController {
             // Path at less than 3 meters from network and route recalculation
 //            origin = Coordinate(coordinate2D: .init(latitude: 48.84458308799957, longitude: 2.3731548097070134), level: 0)
 //            destination = Coordinate(coordinate2D: .init(latitude: 48.84511200990592, longitude: 2.3738383127780676), level: 0)
-            
+
             // Path from level -1 to 0 and route recalculation
 //            origin = Coordinate(coordinate2D: .init(latitude: 48.84445563, longitude: 2.37319782), level: -1)
 //            destination = Coordinate(coordinate2D: .init(latitude: 48.84502948, longitude: 2.37451864), level: 0)
+
+            // Path indoor to outdoor
+//            origin = Coordinate(coordinate2D: .init(latitude: 48.84482873, longitude: 2.37378956), level: 0)
+//            destination = Coordinate(coordinate2D: .init(latitude: 48.8455159, longitude: 2.37305333))
         }
-       
-        map.navigationManager
+        
+        startNavigation(origin: origin, destination: destination)
+    }
+    
+    @IBAction func removeUserCreatedAnnotations() {
+        map.removeAnnotations(userCreatedAnnotations)
+        updateUI()
+    }
+    
+    private func startNavigation(origin: Coordinate?, destination: Coordinate) {
+        disableStartButtons()
+        
+        navigationManager
+//            .startNavigation(from: origin, to: destination, options: Constants.globalNavigationOptions, itinerarySearchOptions: .init(useStairs: false))
             .startNavigation(from: origin, to: destination, options: Constants.globalNavigationOptions)
             .subscribe(
                 onSuccess: { [unowned self] itinerary in
-                    debugPrint("Navigation started successfully")
-                    stopNavigationButton.isEnabled = true
                     simulator?.setItinerary(itinerary)
+                    stopNavigationButton.isEnabled = true
                 },
                 onFailure: { [unowned self] error in
-                    debugPrint("Failed to start navigation with error - \(error)")
-                    updateStartNavigationButtons()
                     ToastHelper.showToast(
-                        message: "Failed to start navigation from - \(origin) to - \(destination) with error - \(error)",
-                        onView: view,
-                        hideDelay: 10
+                        message: "Failed to start navigation from user position to - \(destination) with error - \(error)",
+                        onView: view, hideDelay: Delay.long
                     )
                 }
             ).disposed(by: disposeBag)
     }
     
-    @IBAction func removeUserCreatedAnnotations() {
-        map.removeAnnotations(userCreatedAnnotations)
-        removeUserCreatedAnnotationsButton.isEnabled = false
-        
+    private func updateUI() {
+        startNavigationButton.isEnabled = userCreatedAnnotations.count == 1 && !stopNavigationButton.isEnabled
+        startNavigationFromUserCreatedAnnotationsButton.isEnabled = userCreatedAnnotations.count == 2 && !stopNavigationButton.isEnabled
+        removeUserCreatedAnnotationsButton.isEnabled = !userCreatedAnnotations.isEmpty
+    }
+    
+    private func disableStartButtons() {
         startNavigationButton.isEnabled = false
         startNavigationFromUserCreatedAnnotationsButton.isEnabled = false
     }
     
-    private func updateStartNavigationButtons() {
-        startNavigationButton.isEnabled = userCreatedAnnotations.count == 1 && !stopNavigationButton.isEnabled
-        startNavigationFromUserCreatedAnnotationsButton.isEnabled = userCreatedAnnotations.count == 2 && !stopNavigationButton.isEnabled
+    private func getLevelFromAnnotation(_ annotation: MGLAnnotation) -> [Float] {
+        guard let building = focusedBuilding else {
+            debugPrint("Failed to retrieve focused building. Can't check if annotation is indoor or outdoor")
+            return []
+        }
+        
+        return building.boundingBox.contains(annotation.coordinate) ? [Float(annotation.subtitle!!)!] : []
+    }
+}
+
+extension NavigationViewController: PointOfInterestManagerDelegate {
+    
+    func pointOfInterestManager(_: PointOfInterestManager, didSelectPointOfInterest poi: PointOfInterest) {
+        ToastHelper.showToast(message: "POI clicked with id - \(poi.id)", onView: view, hideDelay: Delay.short) {
+            self.pointOfInterestManager.unselectPOI(poi)
+        }
     }
 }
 
@@ -223,50 +205,35 @@ extension NavigationViewController: NavigationDelegate {
     
     func navigationManager(_: NavigationManager, didUpdateNavigationInfo info: NavigationInfo) {
         navigationInfo.isHidden = false
-        navigationInfo.text = info.shortDescription
+        navigationInfo.text = info.description
     }
     
-    func navigationManager(_: NavigationManager, didStartNavigation _: Itinerary) {
+    func navigationManager(_: NavigationManager, didStartNavigation itinerary: Itinerary) {
+        navigationInfo.isHidden = false
+        ToastHelper.showToast(message: "Navigation started", onView: view)
         stopNavigationButton.isEnabled = true
-        updateStartNavigationButtons()
-        ToastHelper.showToast(
-            message: "Navigation started",
-            onView: view
-        )
+        
+        itinerary.legs.flatMap(\.steps).forEach {
+            let instructions = $0.getNavigationInstructions()
+            debugPrint(instructions)
+        }
     }
     
     func navigationManager(_: NavigationManager, didStopNavigation _: Itinerary) {
+        navigationInfo.isHidden = true
+        ToastHelper.showToast(message: "Navigation stopped", onView: view, hideDelay: Delay.short)
         stopNavigationButton.isEnabled = false
-        updateUIForNavigationStop()
-        ToastHelper.showToast(
-            message: "Navigation stopped",
-            onView: view,
-            hideDelay: 5
-        )
     }
     
     func navigationManager(_: NavigationManager, didArriveAtDestination _: Itinerary) {
-        ToastHelper.showToast(
-            message: "Navigation manager didArriveAtDestination",
-            onView: view,
-            hideDelay: 5,
-            bottomInset: -150
-        )
+        ToastHelper.showToast(message: "Navigation manager didArriveAtDestination", onView: view, hideDelay: Delay.short, bottomInset: Inset.mid)
     }
     
     func navigationManager(_: NavigationManager, didFailWithError error: NavigationError) {
-        ToastHelper.showToast(
-            message: "Navigation failed with error - \(error)",
-            onView: view,
-            hideDelay: 5
-        )
+        ToastHelper.showToast(message: "Navigation failed with error - \(error)", onView: view, hideDelay: Delay.short)
     }
     
     func navigationManager(_: NavigationManager, didRecalculateItinerary itinerary: Itinerary) {
-        ToastHelper.showToast(
-            message: "Navigation itinerary recalculated - \(itinerary)",
-            onView: view,
-            hideDelay: 5
-        )
+        ToastHelper.showToast(message: "Navigation itinerary recalculated - \(itinerary)", onView: view, hideDelay: Delay.short)
     }
 }