+++ Examples/Sources/AppDelegate.swift
@@ -15,6 +15,8 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     
     func application(_: UIApplication, didFinishLaunchingWithOptions _: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
+        SettingsBundleHelper.applySettings(customKeysAndValues: customKeysAndValues())
+
         window = UIWindow(frame: UIScreen.main.bounds)
         
         window?.rootViewController = UIStoryboard(name: "Main", bundle: nil).instantiateInitialViewController()